+++ webknossos/examples/apply_merger_mode.py
@@ -53,7 +53,7 @@ def main() -> None:
     out_layer = output_dataset.add_layer(
         "segmentation_remapped",
         wk.SEGMENTATION_CATEGORY,
-        dtype_per_channel=in_layer.dtype_per_channel,
+        dtype=in_layer.dtype,
         largest_segment_id=in_layer.largest_segment_id,
     )
     out_mag1 = out_layer.add_mag("1")